refactor: move AdaptionUtil to :common:android
For the AnkiActivity move to `:anki-common`

This explicitly adds a dependency ordering: :common:android -> :compat

Any domain-specific extensions to `compat` classes can go into
`:common:android` cleanly

Part of 20737
