+++ AnkiDroid/src/main/java/com/ichi2/anki/utils/AndroidUtils.kt
@@ -31,10 +31,10 @@ import androidx.core.view.WindowInsetsControllerCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import com.ichi2.anki.R
+import com.ichi2.anki.common.android.AdaptionUtil
 import com.ichi2.anki.common.android.appContext
 import com.ichi2.anki.common.utils.android.showThemedToast
 import com.ichi2.anki.snackbar.showSnackbar
-import com.ichi2.utils.AdaptionUtil
 import com.ichi2.utils.copyToClipboard
 import timber.log.Timber
